Remove unused configuration files and update version in composer.json. Deleted .gitignore, .php-cs-fixer.php, mix-manifest.json, and webpack.mix.js. Updated version from 4.32.12 to 4.35.5 in composer.json.

+++ src/Console/InstallCommand.php
@@ -3,8 +3,12 @@
 namespace Laravel\Nova\Console;
 
 use Illuminate\Console\Command;
+use Illuminate\Foundation\Configuration\ApplicationBuilder;
+use Illuminate\Support\ServiceProvider;
 use Illuminate\Support\Str;
+use Symfony\Component\Console\Attribute\AsCommand;
 
+#[AsCommand(name: 'nova:install')]
 class InstallCommand extends Command
 {
     use ResolvesStubPath;
@@ -73,10 +77,6 @@ class InstallCommand extends Command
 
         $appConfig = file_get_contents(config_path('app.php'));
 
-        if (Str::contains($appConfig, "{$namespace}\\Providers\\NovaServiceProvider::class")) {
-            return;
-        }
-
         $lineEndingCount = [
             "\r\n" => substr_count($appConfig, "\r\n"),
             "\r" => substr_count($appConfig, "\r"),
@@ -84,6 +84,30 @@ class InstallCommand extends Command
         ];
 
         $eol = array_keys($lineEndingCount, max($lineEndingCount))[0];
+
+        if (class_exists(ApplicationBuilder::class) && is_file(base_path('bootstrap/providers.php'))) {
+            ServiceProvider::addProviderToBootstrapFile("{$namespace}\\Providers\\NovaServiceProvider");
+
+            if (
+                ! $this->laravel['router']->has('login')
+                && $this->confirm('Would you like to use the Nova login screen as your application\'s default login screen?', true)
+            ) {
+                file_put_contents(
+                    app_path('Providers/NovaServiceProvider.php'),
+                    str_replace(
+                        [$eol.'                ->withAuthenticationRoutes()'.$eol],
+                        [$eol.'                ->withAuthenticationRoutes(default: true)'.$eol],
+                        file_get_contents(app_path('Providers/NovaServiceProvider.php'))
+                    )
+                );
+            }
+
+            return;
+        }
+
+        if (Str::contains($appConfig, "{$namespace}\\Providers\\NovaServiceProvider::class")) {
+            return;
+        }
 
         file_put_contents(config_path('app.php'), str_replace(
             "{$namespace}\\Providers\EventServiceProvider::class,".$eol,
